Refactor repository queries to remove duplication
